+++ doc/conf.py
@@ -56,7 +56,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'fslpy'
-copyright = u'{}, Paul McCarthy, University of Oxford, Oxford, UK'.format(
+copyright = u'{}, FMRIB Centre, University of Oxford, Oxford, UK'.format(
     date.year)
 
 # Links to other things